Quote calculator functionality!! Plus the finished quote page

+++ Intex/Models/Assay.cs
@@ -28,7 +28,7 @@ namespace Intex.Models
 
         [Required]
         [DisplayName("Assay Length")]
-        public double AssayLength { get; set; }
+        public decimal AssayLength { get; set; }
 
     }
 }
